requirements.txt
aws-cdk-lib
constructs
aws-lambda-powertools
boto3
